pyproject.toml
[project]
name = "rag-mcp-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=2.11.3",
    "numpy>=2.3.2",
    "pypdf2>=3.0.1",
    "python-docx>=1.2.0",
    "sentence-transformers>=5.1.0",
    "torch>=2.8.0",
]
